Move withExpectations to the library top package
The reasoning being that this is the entry point and
should be easily discoverable
