+++ core/src/test/scala/eu/monniot/scala3mock/features/ThrowSuite.scala
@@ -1,13 +1,8 @@
 package features
 
-import eu.monniot.scala3mock.main.withExpectations
-import eu.monniot.scala3mock.functions.MockFunctions
-import eu.monniot.scala3mock.main.TestExpectationEx
-import eu.monniot.scala3mock.matchers.MatchAny
+import eu.monniot.scala3mock.ScalaMocks
 
-class ThrowSuite extends munit.FunSuite with MockFunctions {
-
-  def * = new MatchAny
+class ThrowSuite extends munit.FunSuite with ScalaMocks {
 
   case class TestException() extends RuntimeException
   case class AnotherTestException() extends RuntimeException
